Implementada las ventanas gestion empleados y departamentos

+++ AD/Actividad2_JBDC/src/main/java/org/mdv/model/Empleado.java
@@ -14,7 +14,7 @@ public class Empleado {
     private String apellido;
     private String oficio;
     private String fechaAlta;
-    private int salario;
+    private double salario;
     private int comision;
     private int idDepartamento;
